Start implementing ssh verify (WIP)
TODO:
Figure out if parsing an ssh signature into something
that cyrpto.rsa.PublicKey can verify is feasible.

If not, e.g. because a lot of code, or extra signature data, it might
make sense to store the ssh public key in the original RFC4716 format
and make it, and use a custom "ssh-*" scheme and the ssh lib.

Downside: x/ssh is in pre-release.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>